refator Headers and Rows
